refactor: move plugins to modules. Move normalize to utils.
